Almost functional STE tests.
This commit introduces some major issues in regard to testing the STE.

1) What should the test matrices be?
I need to search through the literature and find some specific matrices,
functions, etc. that people are testing STEs with.

2) What are the outputs here? Are they `bool`s?
The Epperly post shows the expectation bounded by epsilon. We can set the
output to `true` and then check for the `<=` residual, but when that
test fails the output conveyed to the programmer isn't fantastic.