Work toward verifying NLA algorithms with non-parametric bootstrap.
Went ahead and ported the `diku-dk/cpprandom uniform_real` distribution;
to sample from the discrete `[a, b]`, we sample from `[0, 1)` and then
scale that up by the range, which should give us a number in the interval.

I've chosen to implement it this way because sampling from the discrete
case requires something like rejection sampling, which if we don't have
a state (we have a counter) isn't the cleanest solution...

That solution might look like advancing the `key` value for each rejected
sample, but I digress.