Chebyshev tests are still busted.
As `t` becomes smaller, the tests are more likely to fail. The Hutchinson tests
that *should* produce the correct result everytime (because the numerics are solely
on the diagonal) fail. For these matrices, we don't particularly need a Chebyshev
bound. It'd be smart to just check for equivalence, or near equivalence.