Gaussian embeddings.
I don't know if these should be in the `sketch` file; I've read an NLA manuscript
that takes the approach that these operations are "sketch" operations, such as
"left sketch", right "sketch", etc.

The argument to place this in a seperate `embedding` module is that the
Gaussian distribution parameters are known and dependent on the embedding dimension.

I don't know -- we'll just push this at this point. I need to build stuff around
Sketch-and-Solve and Sketch-and-Precondition.