A naive subspace iteration.
If we subspace iterate, we lower the spectral decay, making
the rangefinder more accurate.

I don't know the preferred API, but I'm going to throw this
in here. This is also the naive version -- the preferred method
is going to involve either computing the QR after a few iterations
or Krylov subspaces.