Work on a TSQR implementation.
So here's the rub: we can't use a `reduce` operator because the
`R` matrices are blocked on top of each other at the end of each
reduction operation (the `a -> a -> a` type won't hold; see
https://bebop.cs.berkeley.edu/pubs/mhoemmen2008-tsqr-tech-report.pdf)

Can we use a `loop` operator instead and then do `flatten` / `unflatten`
to what we desire?

Of note: the row value of the input matrix must be divisible by `k`,
and `k` itself must be a power of 2 in order for the tree structure.