Work on a bunch of things: rangefider, svd, tsqr.
Modularize the TSQR implementation. Tro22 recommends to not use GS;
unfortunately the Householder implementation within `linalg` is a
blocked Householder. We need to find an economical QR.

We can likely modularize the rangefinder; the `sketch` can be parameterized
and passed in with some test matrix. Additionally, there's some potential
to do a subspace iteration.

The API of `sketch` is becoming problematic...